LABS/LAB DOCKER/backend/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
psycopg2-binary==2.9.9
